bin: implement autospawning of the server
Users may find it convenient to perform an operation the same way
whether they're working on their local machine or on a remote machine.
In order to make this work, let's autospawn a server, so one is always
present on the local machine.

Check for and honor a value called "root" in the configuration file to
determine whether the machine we're on is local or remote.  As a default
value, check for the lack of SSH_TTY, plus something indicating we have
a display.  This default may change in the future.

Adjust the server so that its run functions don't return until the
server is started, which lets us avoid a race condition trying to
connect to an unbound socket even though the server is starting up.

Move the --no-detach and --socket options to be root options since now
any program can use them.  Search for and discover any sockets in any of
the directories and pick the first one that can be connected to.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>