bin: implement autopruning sockets
Currently, if we have a stale socket but no server running, we'll fail
to connect, since the socket is already in use and the server cannot be
started.  Let's make this a little easier by implementing optional
autopruning of stale sockets: if we find a socket we cannot connect to,
we unlink it, and then we'll be able to start the new socket.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>