bin/template: implement single-quoting
Environment variables can usually contain arbitrary non-NUL bytes.
Let's make it easy for users to write secure and functional
configuration by allowing them to automatically shell-quote another
value.

Since we know Git does this properly (and is in fact the inspiration for
this technique), use output from it as part of our testcase.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>