lawn-bin: add an SSH proxy
While it's possible to forward a socket with OpenSSH, this has several
downsides.  First, it's impossible to forward a socket into the
appropriate /run/user directory unless you know the remote user's user
ID, because OpenSSH doesn't expose this.  Second, OpenSSH will just
silently fail if the destination socket already exists, meaning that as
a practical matter, it's only safe to forward sockets with a timestamp
or random component, which is ugly.

Instead, let's add a proxy which can tunnel messages over the SSH agent
socket.  There are both client and server components to allow proxying
these messages.

Because the SSH agent protocol is such that the server only speaks when
spoken to, this unfortunately requires polling, although that might be
acceptable in many cases.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>