lawn-bin: allow using real booleans for boolean values
It seems silly to invoke a shell when the user really wants a simple
boolean.  In such a case, allow the user to specify the boolean as a
normal YAML boolean.

This also allows us in the future to specify strings that are either
shell expressions or literal strings as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>