Start a 9P server
When working with ephemeral development environments, it's often useful
to save things like shell history.  In order to do that, let's implement
a 9P server to offer a place to save these files.  The Linux kernel can
mount such a server, providing the ability to save those files.

There are three variants of this protocol: the original 9P2000, the Unix
variant 9P2000.u, and the Linux-specific variant 9P2000.L.  We implement
all three using basically the same interface and distinguish between
them in the case we need to by using a metadata field.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>