lawn-constants: implement several traits for Error
In order to make the Error type more usable, implement several traits.
Provide Error, and with it Display and Debug, since this is designed as
an error type.  Implement Copy and Clone, since it's essentially an
integer.  Implement Hash so it can be used in a HashMap or HashSet.
Finally, implement traits providing for comparisons.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>