lawn-bin: avoid extremely high-rate channel polling
Right now, we just poll channels in a loop in the client.  That works
fine if the poll operation is slow, but it's very noisy if the channel
is fast to get back to us.  In such a case, let's poll the channel using
stepped timing unless we know that there's data waiting or some recent
activity on the channel.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>