lawn-bin: read more efficient sizes
When we perform a test read, we read only a single byte.  We then
process the file descriptor.  However, if there is any data, even a
single byte, we don't perform any other reads and handle only the data
that's there.  Now that we're polling less often, this can read to a
large number of single byte reads, which are inefficient.

Instead, let's read a decent-sized chunk and then process it
immediately, and if there are more things to process, we can then get to
them in the next round.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>