lawn-bin: move polling out into a separate function
Currently, polling is in the struct specific to a generic command
channel.  Let's move it out so that we can use the same code to poll
anything with file descriptors regardless of the kind of channel.  This
will be useful when we implement a 9P channel, which does not run a
command and won't use the generic command channel.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>