lawn-constants: add AsLogStr trait
In the 9P protocol, as elsewhere, we end up with a large number of byte
strings.  It would be helpful to log them in a consistent way so as to
print them as ASCII text when possible and escape them when it's not.
Add a trait, AsLogStr, which can be used to format byte strings in this
way.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>