lawn-bin: add configuration for 9P server
In the future, we'll implement a 9P server.  To do so, we'll need to
configure it so it can be used.  Let's add some configuration for it.
Since we can't create an identifier starting with a digit, use "p9p",
where we just repeat the "p" for protocol at the beginning.

Like our other configuration, let's add a condition, and also allow the
user to specify a command so they can script this value (e.g., to
provide a per-host directory).

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>