lawn-9p: implement a libc-based 9P backend
In lawn, our primary goal with implementing 9P is to allow the remote
system to access a limited amount of the local file system.  To that
end, let's implement a backend which just calls local file system
operations.

We specifically prevent the remote user from escaping the root, since
the local user will have specified a specific directory to use.
Attempting to escape it will result in an EACCES.

We also attempt when possible to use file descriptors for operations to
provide more predictable behaviour.  It may be in the future that we
prefer a purely path-based approach for consistency, but this is what we
have for now.

Make sure that we have the main library and the Cargo files so that we
can build the project successfully.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>