lawn-bin: create a 9P channel
In order to speak 9P over our lawn socket, we'll need a 9P channel.
Create one for the server side, and use a Unix socket pair for
communication instead of duplex channel, since the sockets can be polled
in a consistent way.

Implement an authenticator as well, which is required.  Use the values
we get from creating the channel to populate the authenticator and
ignore the values that the client has passed, since every separate 9P
mount will have its own socket.  In general, most 9P clients don't
provide useful values here except possibly for the username and user ID,
so ignoring them is the safe bet.

Note that when we try to perform a read or write and fail because we're
nonblocking, we must manually convert the io::ErrorKind::WouldBlock into
EAGAIN, since the former is created by Tokio and doesn't come from
calling a C function.  If we didn't do that, the lack of an error code
would mean we'd map the result to EINVAL, which would result in broken
behaviour.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>