lawn-bin: allow the server to create 9P channels
Now that we have a 9P channel available, let's let the server create
them if they're available and one is requested.

Note that presently this code is completely inaccessible, since we never
send the 9P channel capability and therefore nobody can negotiate that.
However, in the future, we will do so, and then our channel will
automatically work.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>