lawn-bin: add support for mounting 9P connections
We now have a 9P server, but we need to accept connections.  Add a
struct to proxy data from a 9P socket over the lawn socket and back.
Add a subcommand, mount, which starts this proxy and runs one or more
commands with either a socket or the standard input and output wired up
to the 9P operations.  The former allows the use of 9pfuse, and the
latter allows mounting with the Linux kernel, which cannot use a Unix
socket for this purpose.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>