lawn-9p: avoid opening files needlessly
In some cases, we may end up with a file or directory which is
unreadable.  In such a case, our current approach of opening every file
and directory is untenable, since it prevents us from working with these
files.  It may also exhaust resources on systems where the number of
file descriptors is limited, such as the default on macOS, which is 256.

Instead, let's move from a large number of types of file IDs to three:
an auth descriptor, an open file, and a closed file.  Since an open file
can be a file or directory, we now keep track of this by stating the
file if necessary.  If a file is not open, it has no file descriptor
associated with it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>