lawn-9p: fix reading directories
The current code for reading directories has a small problem: if an
entry is read but is too large for the buffer, then we'll end up
dropping that entry because the iterator won't read it next time we go
through.

To solve this, let's use a Peekable iterator to look at the next entry
and see if it's too big.  This requires, however, that we know fully up
front how big the entry is, which means that for the Unix version of the
protocol, we know the value of the extension field: either the
destination of the symlink or the block or character device info.

This is a bunch of expensive information, so in the case when we're
normally going through the loop and computing the next entry to see if
it's too big, let's simply save that computation for the next time so we
don't need to call lstat or readlink multiple times.  Similarly, since
we're calling those functions up front when reading entries, don't call
them again once we're serializing those entries.  Instead, save
everything that we need for later on.

Add some tests for each of the variants of the protocol using an
intentionally small buffer size so we can verify we're processing data
correctly and that we're getting all of the entries we expect.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>