lawn-9p: don't try to open final walk component
If we're walking multiple path components, we need to open each
intermediate directory to traverse the paths securely.  However, for the
final component, we don't need to open it, and it could well be
something that we'll never be able to open, such as a Unix socket.
Let's make our code both simpler and more robust by not opening it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>