Improve polling performance
Right now, our polling code doesn't let us register a list of events
we're interested in.  As a result, for a 9P channel, where the file
descriptor corresponding to selector 1 is open for both input and
output, we can end up with spurious wakeups because we register an
interest in input (which is practically always ready) but we really only
care about output (which is only ready when the 9P server has sent a
message).

Let's provide an optional extension to the message to indicate what
events we're interested in.  This in fact results in substantially
better performance, possibly by an order of magnitude.  For example,
when trying to clone Git's repository without this change, the operation
has not even gotten to the point of sending messages from the remote by
1:19.  However, with this change, the entire repository is cloned and
checked out in only 1:44.  The performance using a local lawn server is
so substantially improved that in some cases it feels like it performs
as a local disk.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>