lawn-bin: remove hard-coded check for /run/user
This hard-coded check was originally designed to gracefully handle the
fact that XDG_RUNTIME_DIR is usually not set without a desktop
environment.  However, using a hard-coded path makes it difficult to
test because we can't adjust the environment variables to configure
socket discovery.  As a consequence, remove this socket location.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>