lawn-bin: fix autospawning
Lawn is supposed to automatically start a server if one is not running
and we're on the root (local) machine.  However, this had stopped
working because we attempted to spawn a Tokio runtime within a runtime.
The second runtime is the one created in the forked child for the server
itself, and we cannot avoid creating this runtime.  The fork will have
terminated all but the current thread, so any runtime creation (which
spawns threads) must happen after the fork.

However, we have some code which is async which must run as part of the
socket autodetection process.  For example, the SSH proxy code is async
and we must run it within a runtime, plus the code to perform the client
probe.  Pass a handle to the runtime to allow the relevant types to be
created with an EnterGuard, plus spawning the actual server and running
the probe in the appropriate async blocks.  This fixes our autospawning.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>