lawn-bin: update client to handle EAGAIN gracefully
Right now, if we're draining the pipe from when the child process has
ended, we can sometimes get EAGAIN due to our relatively short timeout
of 10 ms.  We can actually get an EAGAIN at anytime, but we only see
this as a problem when draining because we won't retry reading again.

Let's make sure that we read the data for the final drain until we
actually get EOF.  As part of that, ignore any cached data that is
empty, and if we read zero bytes, consider that EOF and close the
selector.  This means that we won't end up with truncated data when the
read operation is slow.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>