lawn-9p: handle FreeBSD usage of EMLINK with O_NOFOLLOW
In case a symlink is encountered when using open with O_NOFOLLOW, POSIX
mandates an error of ELOOP.  FreeBSD instead returns EMLINK, so let's
handle that error as well.  Note that while NetBSD also returns a
nonstandard error code (EFTYPE), we map that to ELOOP internally, so
this is sufficient.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>