lawn-bin: use command instead of hard-coding a path
On FreeBSD, true is in /usr/bin, not /bin.  As a result, an attempt to
use the return value to see if a condition is true fails, since an exit
status of 127 is not a success.  To avoid this problem, use `command
true` and `command false` instead of specifying a hard-coded path.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>