lawn-9p: fix situation where root contains symlink
Currently, all the tests fail on macOS because the temporary directory,
and hence the root for all of our tests, is a symlink.  Use the
canonicalized version of the path when we write to `self.root`, and
adjust the tests to verify against a canonicalized version of the root
as well.

Mark the dir member as allowed to be unused.  We must maintain the
object because the temporary directory holding our working data will be
deleted when it is dropped.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>