lawn-bin: avoid spinning when no standard input
When the process has exited and there's nothing to read from the
standard input, that's fine: we don't need to keep trying to read that
data.  Mark this case as not having seen any data the last time and skip
processing standard input that didn't have data the last time.  After
all, if the process has exited, we can't send it to the server anyway.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>