lawn-bin: fix mounting when using SSH socket
When using an SSH socket, the socket we get won't have a pathname, so
currently the mount command panics.  Let's fix this by detecting this
case and spawning a Proxy instead of a ProxyListener that can reuse a
connection instead of spawning a new connection.  As part of this, make
the Connection into an Arc so we can clone it easily.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>