lawn-constants: avoid formatting text when logging is disabled
The conditions in the lawn-constants logging macros were backwards,
leading to the situation where at a normal logging level, all of the
trace logging was formatted needlessly.  This can consume a substantial
amount of resources in heavy usage (about 8%), so fix it to improve
performance.  Note that we put the tests in lawn-bin because the macros
in lawn-constants refer directly to the root crate name, which means
that the macros don't work from within the crate.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>