lawn-9p: don't rewrite absolute symlinks
If we provide an absolute symlink and the remote kernel reads it, it
will rewrite it into a path outside of our mount if that's appropriate.
That's fine, and rewriting absolute symlinks here results in a bizarre
situation where the value read differs from what was written since we
don't rewrite the equivalents to readlink.  Avoid producing weird,
invalid symlinks by disabling the rewriting.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>