lawn-9p: allow removing broken symlinks
When we remove an FID, we need to call different functions depending on
whether the path is a directory or not.  In order to determine that, we
stat the file.  However, we used the `fs::metadata` function, not the
`fs::symlink_metadata` function, so if we were pointing to a broken
symlink, then we would get ENOENT and be unable to remove the file.  Use
the lstat equivalent instead to allow removing broken symlinks.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>