Add some RSpec integration tests
While the built-in Rust tests are helpful for testing lots of things, we
need some integration tests.  Let's write these in RSpec, since we can
write these easily by scripting input and output.  Add a fake xsel
binary as well, which can mimic the real xsel, so we can write some
clipboard tests.

Because the version of Ruby on NetBSD is now 3.1, upgrade the use of
Asciidoctor to 3.1 as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>