lawn-bin: allow running standard commands
In most cases, we want to operate with a Tokio command, since most of
our code is async.  However, Tokio commands lack one important feature:
their pipes cannot be converted _into_ a raw file descriptor, only
borrowed _as_ a raw file descriptor.

This is important because when converting the process pipes into
PipeRead and PipeWrite from tokio-pipe, we need to make sure the
descriptors are not already part of the runtime.  If they are, we'll end
up with an EEXIST error and a panic.  As a result, add helpers that
allow us to spawn a standard Command instead of a Tokio command.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>