lawn-bin: extract proper error
Currently, we never detect that the remote side has sent us an Errno
value because we unpack the incorrect type.  We try to extract a
protocol error, but that's always wrapped in a handler error first.  As
a result, we don't properly detect the EAGAIN status and can mishandle
it.  Fix this by extracting the correct type.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>