lawn-bin: switch to using tokio-pipe for commands
Right now, we convert our child standard input, output, and error into
files and use those for interacting with the process.  This lets us
treat all of our readers and writers the same way, which is important
for efficient implementation.

However, because Tokio uses blocking I/O for its files but not its
pipes, we can end up with cases where we get EAGAIN if the pipe is full,
which Tokio doesn't expect.  The problem then becomes that Tokio loses
the data we write next to the pipe, even though it says it has
succeeded.

Convert to using a standard command instead of a Tokio command, since we
need to extract and take ownership of the file descriptors with
IntoRawFd, and Tokio doesn't let us do that with its command pipes.
Convert those into an appropriate tokio-pipe object so we can work with
them as we want.

Force our file descriptors into an Arc and use a Tokio mutex to process
them.  This is necessary because when processing the file descriptors
out of the tuple, the read and write descriptors are no longer of the
same type, and thus when extracting the file descriptors, we need to
lock the inner mutex.  This must be done by passing the entire set of
descriptors into the async block, which in turn means that we must make
the RwLock async or the future won't be send.  For similar reasons, they
must now live in an Arc so they will live long enough for the async
block to complete.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>