spec: add tests for commands
We have tests for clipboard functionality, but not commands.  Much of
the infrastructure is shared internally between the two, but to better
verify, let's add some tests that cover basic functionality by
implementing a small sha256sum script.

Include tests that verify that standard input is closed properly, which
our sha256sum requires to work, and also add some tests that insert
small pauses so that we can verify we handle buffering and delays
gracefully.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>