Provide READMEs for each crate
We want to provide a Markdown README for each crate, since that's the
format that crates.io knows how to render properly.  Since we prefer
AsciiDoc, let's end up providing some Makefile targets to generate the
Markdown from AsciiDoc by converting first to DocBook and then using
pandoc to convert to Markdown.  Note that the conversion to DocBook is
necessary because pandoc doesn't know how to parse AsciiDoc.

Use the include features of AsciiDoc to use a template for most of the
content and only add in the extra content necessary for each target.
For the main lawn crate, use the main README, but fix up the links so it
works when it's uploaded.  Ignore this generated file so we don't
accidentally check it in.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>