lawn: expose user-agent information over the protocol
It's possible we may have multiple implementations of the protocol, and
even if we don't, we may want to know what version is on the other end
for debugging and informational purposes.

To make this possible, expose an optional user-agent which can be sent
over the protocol to inform the other side what implementation and
version we are.  This is sent in the capability response and the
version request.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>