.github/workflows: switch to ruby/setup-ruby
GitHub Actions warns that Node.js 12 actions are deprecated.  In order
to upgrade to Node.js 16 actions, we need to switch from
actions/setup-ruby, which is deprecated and abandoned, to
ruby/setup-ruby, which is actively maintained.

We must also now specify a Ruby version, so do that as well.  We use
3.1, but the exact version doesn't matter very much.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>