spec: implement debugging
It can be helpful when debugging to have the tests print some debug
information.  Let's make it easy to do so with an environment variable
and some helpers that developers can use to clean up and print data
early.

Move the entire server data structure into the remover, not just the
PID, so we can read the error file and print its contents.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>