spec: wait for server socket
In some cases, the `lawn server` process startup can take some time, and
by the time we return from the `start_server` function, the socket may
not be in place.  This can most notably seen when running in emulation,
as we do in CI.

If the socket is not in place, we may fail to run other commands, since
in our tests, we typically don't meet the criteria for a root machine
and so we won't autospawn a server.  In that case, any commands which
run after that will fail spuriously, which results in flaky tests.

If the socket is absent after starting the server, wait until it's
present before continuing to avoid failing tests.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>