server: fix channel error handling
We currently log errors when command condition and pre-hooks checks
succeed.  Those aren't really errors and they tend to show up on the
screen even when the server is running with normal verbosity.

Instead, let's flag actual errors as errors in all channel operations,
and turn successful operations to debug.  We're not interested in
tracing the actual flow of operations, so trace is not appropriate, but
debug is appropriate for logging successful operations that occur in the
normal course of running the program.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>