lawn: process messages in a task
`ProtocolHandler::recv` is not cancellation safe because it expects to
hold the lock for the duration of the read, and the read can be
interrupted and restarted, meaning the lock is reacquired.  We need to
ensure that the `recv` call is not present in the `select!` body to make
sure that it isn't accidentally canceled.

To do this, move it into a loop and use a channel to return the value
that can be used in a select.  This will be cancellation safe, and is
what we do in the client as well.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>