lawn-protocol: add types for allocation extension ranges
Right now, we support extensions with the syntax `extension@domain` but
don't provide a way of handling protocol-specific messages.  Doing so is
a helpful way to allow others to extend the protocol, so let's implement
it.

We define protocol extension messages dynamically with message types and
response codes with 0xff as the top byte.  If the implementation
supports the `extension=allocate` feature, then we can use the provided
messages to dynamically allocate a range of codes for the extension to
use.

Note that an extension is not obligated to use these codes.  For
example, an extension which creates a new type of channel will probably
use the standard channel codes to implement that functionality.
However, they are available for implementations which wish to use them.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>