Implement suitable messages for continuations
Right now, we have a simple request-response model.  However, some
operations, like some SASL authentication mechanisms or paginated
requests, may need additional data to continue or may return partial
results.

Let's fully flesh out our continuation support.  The client may send a
request and get a `Continuation` response instead of a `Success`
response.  In that case, it should either try again with the `Continue`
request, plus the kind and ID of the previous request, or it should
abort the partial request with the `Abort` request and the kind and ID
of the previous request.

Aborting a partial request allows the server to free resources.  For
example, if a client requests a response which is so large as to be
paginated, but doesn't care about more than the first few items, it can
abort the request to signal its intentions to the server.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>