server: allow sending continuation responses
Right now, the server's `process_message` response returns a boolean to
indicate whether the connection should be closed.  However, with
continuation messages, there's really three types of successful
responses we can give: a successful response that indicates we'd like to
close the connection, a successful response that indicates we don't want
to close the connection, or a continuation response.

Since we now have three options, let's make our boolean into an enum and
return values accordingly.  If we get a continuation message, let's send
an appropriate response using the handler.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>