Create a builder for Config
Right now, it's hard to add more types of configuration to Config
because it takes so many arguments.  Let's make this easier by creating
a builder, which we can then use in our tests.

In the future, we'll want to control which capabilities are advertised
in our server, and we'll be able to do that by adding another option to
the builder.

As part of this, instantiate the configuration object at the time of
building our test instance and use it throughout the instance.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>