lawn: implement continuations as another success
Right now, we have two kinds of successful status codes: Success (0x0)
and Continuation (0x1).  However, our serializer code expects only one
kind of message to deserialize, and typically a continuation will have
different data than a simple success.

To deal with this, let's create a ResponseValue enum that deals with
either a simple success or a continuation, and use it wherever we
deserialize data.  Update the client and server to take this type, and
to return a suitable error if a continuation is received and unexpected.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>