Allow arbitrary capability types
We currently have a small, well-known set of capabilities.  However,
there might be a large number of vendor-specified capabilities and we'd
like to be able to represent those, even if we don't intend to support
them.  Create an `Other` variant that encodes the capability in the
value.

Since all capabilities are now represented, implement `From` instead of
`TryFrom`, and since our capabilities no longer can be represented with
static byte strings, get rid of those variants as well.  Similarly,
since Bytes is not `Copy`, remove the implementation of that trait.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>