error: implement `TryFrom` for various error types
In many cases, we have a handler error or a protocol error that we'd
like to extract from a generic error, such as in the client code.  Let's
make our lives easier by implementing `TryFrom` for these types to see
if we can unwrap the inner types more simply and easily.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>