Implement mapping of extensions
Right now, we have a list of known capabilities.  However, it's possible
that people may want to extend the protocol with third-party
capabilities that have custom messages.  Without a way to allocate
custom messages, this would require users to request allocations from
the project.

Instead, let's implement a way to dynamically allocate message and
response codes.  A user may request a range of codes where the top byte
is 0xff that correspond to an extension.  Once that range has been
allocated, a user may use these codes both as request and response codes
until the extension range is deregistered.

Additionally, provide a way to list mapped extensions.  Since this is a
list command and the response may be large, implement a suitable
pagination mechanism using continuations to send a large number of
responses.

Note that currently each extension allocates a range of 4096 elements
(12 bits) and up to 4096 ranges (another 12 bits) may be allocated.
This may change in the future, but is expected to be sufficient to allow
most types of extensions in the future.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>