Don't require auth for extension allocation
The protocol documentation specifically anticipates that some
capabilities may modify the protocol.  For example, a capability might
offer an encrypted tunnel like the STARTTLS command in many IETF
protocols.

If this is from a built-in Lawn message, then this poses no problem.
However, if it's an extension, this does, since authentication is
currently required in order to create and delete extension requests.
Perhaps the user is using something other than a Unix socket or SSH
agent connection over an SSH tunnel and does not want to send any
authentication credentials over an unencrypted connection.

Since these messages are not unlike the capability and version messages,
let's allow them to be used without authentication.

Note that we still require authentication for listing extensions,
because this may involve a continuation and keeping some server-side
state.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>