tty: add definitions for TTY settings
In order to set up a TTY, we need to set up a definition of TTY
settings, since different TTYs work differently.  Provide the encoding
of the values that are used by RFC 4254 and RFC 8610, which we
incorporate here.  Add a structure to control TTY settings that can be
used to read and write the values and serialize them to a map that we
can send as part of a message.

Include `VMIN` and `VTIME`, which are not exposed over the protocol,
since we need them to configure the TTY properly in raw mode.

The set of raw file settings comes from OpenSSH and Dropbear, which use
this for setting terminal modes in the SSH protocol.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>