doc: fix installation instructions
The current installation instructions assume that the user has cloned
the repository.  Having uploaded the crate to crates.io, we don't need
to require people to clone the repository anymore.  Update the
directions so that people can use the version from crates.io
automatically, which should simplify installation quite a bit.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>