client: use intersection for capabilities
In set theory, the union of two sets contains those items which are in
either set; that is, it's equivalent to a logical OR.  The intersection
of two sets contains the items in both sets; that is, it's equivalent to
a logical AND.

When negotiating capabilities between the client and the server, make
the client negotiate those capabilities in the intersection; that is,
only those capabilities that both sides support.  Right now, we
negotiate the union, which the server will reject if we propose any
capabilities it doesn't have.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>