Add basic types for generic file system access
Right now, we have support for 9P.  However, we'd also like to add
support for SFTP in the future.  Let's add some basic types, including a
`Backend` trait and supporting types, that can be used to implement a
generic file system interface capable of supporting either protocol.

Note that we specifically define types for protocols here.  This is
because we want backends to be able to optimize the information given.
For example, the `QID` type field is irrelevant in SFTP, so we want to
be able to skip any `stat` calls that might be necessary to fill it in
if we don't need it.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>