lawn-fs: add a libc backend
We'd like to make our current 9P backend a little more generic for
dealing with both 9P and SFTP, as well as a variety of other protocols.
Let's add a backend which is similar to the 9P backend, but with a few
changes.

First, let's remove some of the older 9P functionality that's
duplicative, leaving only the more standard Unix functionality.
Additionally, let's avoid operating in place on FIDs, instead returning
new ones when possible.  Make the readdir code more generic to allow
working with any of the 9P protocols without needing to know which one
is in use.  Finally, add support for the `resolve` and `realpath`
methods, which are useful in SFTP and which are not efficiently
implemented using the existing methods.

We also allow the user to indicate that they don't need a valid QID,
which allows us to avoid many expensive stats.

Note that this backend uses rustix to paper over many of the 64-bit
weirdnesses between machines and produce nicer, simpler code.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>