lawn: add support for SFTP channels
We now have a fully-functional SFTP server.  Let's wire it up as a
suitable channel via the configuration and channel code.

Note that as part of this, we define a generic configuration category
for file systems, `fs`, that can be used by 9P or SFTP.  If there is a
9P mount with the same name, that will override the generic mount for
9P, but only the generic mount is used for SFTP.

Intentionally try to share as much code as possible for SFTP and 9P
channels, since they are very similar and should therefore function
similarly.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>