lawn: wire up SFTP support
Our SFTP support is now functional.  Let's wire it up in the main
program, calling into the appropriate channels.  Turn the 9P proxy into
a generic proxy that can handle either 9P or SFTP.  Add a new `--type`
argument to `lawn mount` to specify the type of mount to use.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>