clippy: disable bad bit mask warnings
`lawn_fs::backend::O_RDONLY` is zero (because Linux has that value as
zero) and that causes the `bitflags` crate to potentially perform a
bitwise-and against zero, which triggers this warning.  We need this
value to be configured, but we can't disable this warning in the
`lawn-fs` crate itself because the warning comes from another crate.
Disable it in the Makefile instead.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>