mount: add support for automatically mounting a mount
With SFTP mounts, we know that usually the user will want to mount using
SSHFS.  To make their life easier and reduce the need to look at the
documentation, let's add an `--auto` option which automatically invokes
the `sshfs` binary with the proper args.

In addition, let's update the getting started documentation to suggest
this approach instead, as well as the documentation for `lawn mount`.

Signed-off-by: brian m. carlson <sandals@crustytoothpaste.net>